Move some battle code out of S_ITEM
